Move HostUtils to the common package
